Added edit Profile icon to different fields
